Remove Air page from dashboard
