+++ edscrapers/tools/dashboard/json_parser.py
@@ -88,3 +88,16 @@ def get_table_rows_by_office(key, is_sorted=True):
 def get_intersection_data():
     data = read_json_file()
     return data['intersections']
+
+def get_datasets_bars_data(is_sorted=True):
+
+    data_list = list()
+    total_res_dict = get_total_datasets_data()
+    for key, value in total_res_dict.items():
+        data_list.append({
+            'x': ['Datasets'], 'y': [value],
+            'type': 'bar', 'name': key
+        })
+    if is_sorted is True:
+        data_list.sort(key=lambda item: item['y'][0], reverse=True)
+    return data_list
